Changed WLAN behaviour
The Scale will now at startup try to connect to a known WLAN (which can
be set in the configuration).
If the WLAN is not found or we can't connect within 10s to the WLAN a
accesspoint will be created and the GUI can be accessed via cgscale.local
(default).
This is also the initial behaviour after flashing the firmware. In
accesspoint mode the scale can be used normally or the WLAN can be
configured.